refactor(doctor): split tags.go and remediate.go into focused files
Extract code from tags.go (772→335 lines) and remediate.go (990→400 lines)
into three new files, all staying in package doctor:

- version.go: semver parsing, version comparison (pure computation, zero I/O)
- tagging.go: tag suggestions, picker curation, reorderSuggestions
- apply.go: file mutation (applyPin, applySHAToTag, applyReResolve),
  ref approval tracking, and helper functions

No logic changes — pure structural refactor.
