+++ internal/doctor/tags.go
@@ -5,10 +5,7 @@ import (
 	"net/url"
 	"os"
 	"path/filepath"
-	"regexp"
 	"sort"
-	"strconv"
-	"strings"
 	"time"
 
 	"github.com/cli/go-gh/v2/pkg/api"
@@ -22,205 +19,6 @@ type TagInfo struct {
 	IsRelease   bool   // true if a GitHub Release exists for this tag
 	IsImmutable bool   // true if the release is marked immutable (tag can't be moved/deleted)
 	IsMajor     bool   // true if this looks like a major-only tag (e.g. "v4")
-}
-
-// semver holds parsed version components.
-type semver struct {
-	Prefix string // "v" or ""
-	Major  int
-	Minor  int
-	Patch  int
-	Rest   string // anything after patch (e.g. "-beta.1")
-	Raw    string
-}
-
-var semverRE = regexp.MustCompile(`^(v?)(\d+)(?:\.(\d+))?(?:\.(\d+))?(.*)$`)
-
-func parseSemver(tag string) (semver, bool) {
-	m := semverRE.FindStringSubmatch(tag)
-	if m == nil {
-		return semver{}, false
-	}
-	major, _ := strconv.Atoi(m[2])
-	minor := 0
-	if m[3] != "" {
-		minor, _ = strconv.Atoi(m[3])
-	}
-	patch := 0
-	if m[4] != "" {
-		patch, _ = strconv.Atoi(m[4])
-	}
-	return semver{
-		Prefix: m[1],
-		Major:  major,
-		Minor:  minor,
-		Patch:  patch,
-		Rest:   m[5],
-		Raw:    tag,
-	}, true
-}
-
-func (s semver) MajorTag() string { return fmt.Sprintf("%s%d", s.Prefix, s.Major) }
-func (s semver) MinorTag() string { return fmt.Sprintf("%s%d.%d", s.Prefix, s.Major, s.Minor) }
-
-// IsFullSemver returns true if the version has all three components (major.minor.patch)
-// and no pre-release suffix. Tags like "v4" or "v4.2" return false.
-func (s semver) IsFullSemver() bool {
-	return s.Rest == "" && s.Raw != s.MajorTag() && s.Raw != s.MinorTag()
-}
-
-// IsMutableVersionTag returns true if ref looks like a mutable version tag
-// (major-only like "v4" or minor-only like "v4.2") that should be narrowed
-// to a specific patch version for pinning.
-func IsMutableVersionTag(ref string) bool {
-	sv, ok := parseSemver(ref)
-	if !ok {
-		return false
-	}
-	return !sv.IsFullSemver()
-}
-
-// IsNarrowedVersion returns true if narrowed is a more specific patch version
-// of mutable. For example: mutable="v4", narrowed="v4.1.0" → true.
-// mutable="v4.2", narrowed="v4.2.1" → true. mutable="v4", narrowed="v5.0.0" → false.
-func IsNarrowedVersion(mutable, narrowed string) bool {
-	mv, mOK := parseSemver(mutable)
-	nv, nOK := parseSemver(narrowed)
-	if !mOK || !nOK {
-		return false
-	}
-	if !nv.IsFullSemver() {
-		return false
-	}
-	// Major must match.
-	if mv.Major != nv.Major {
-		return false
-	}
-	// If mutable specifies minor (e.g. "v4.2"), narrowed minor must match.
-	if mutable != mv.MajorTag() && mv.Minor != nv.Minor {
-		return false
-	}
-	return true
-}
-
-// BestPatchTagForSHA returns the highest full-semver patch tag pointing at the
-// given SHA, or "" if none exists. This is used to narrow mutable version refs
-// (like "v4") to a specific patch version (like "v4.2.1") when pinning.
-func (tl *TagLister) BestPatchTagForSHA(owner, repo, sha string) (string, error) {
-	matching, err := tl.TagsForSHA(owner, repo, sha)
-	if err != nil {
-		return "", err
-	}
-
-	var best semver
-	bestFound := false
-	for _, t := range matching {
-		if t.IsMajor {
-			continue
-		}
-		sv, ok := parseSemver(t.Name)
-		if !ok || !sv.IsFullSemver() {
-			continue
-		}
-		if !bestFound || sv.Major > best.Major ||
-			(sv.Major == best.Major && sv.Minor > best.Minor) ||
-			(sv.Major == best.Major && sv.Minor == best.Minor && sv.Patch > best.Patch) {
-			best = sv
-			bestFound = true
-		}
-	}
-
-	if !bestFound {
-		return "", nil
-	}
-	return best.Raw, nil
-}
-
-// UniquePatchTagForRef returns the sole full-semver patch tag that matches the
-// given ref's family, or "" if the choice is ambiguous (0 or 2+ candidates).
-// For "v9" it only considers v9.x.y tags; for "v4.2" only v4.2.x tags.
-// This is used for auto-pinning: if there's exactly one obvious patch tag,
-// we can pin without prompting.
-func (tl *TagLister) UniquePatchTagForRef(owner, repo, sha, ref string) (string, error) {
-	refSV, refOK := parseSemver(ref)
-	if !refOK {
-		return "", nil
-	}
-
-	matching, err := tl.TagsForSHA(owner, repo, sha)
-	if err != nil {
-		return "", err
-	}
-
-	var candidates []semver
-	for _, t := range matching {
-		if t.IsMajor {
-			continue
-		}
-		sv, ok := parseSemver(t.Name)
-		if !ok || !sv.IsFullSemver() {
-			continue
-		}
-		// Must be in the same family as the original ref.
-		if sv.Major != refSV.Major {
-			continue
-		}
-		// If original ref specifies minor (e.g. "v4.2"), patch must match that minor.
-		if ref != refSV.MajorTag() && sv.Minor != refSV.Minor {
-			continue
-		}
-		candidates = append(candidates, sv)
-	}
-
-	if len(candidates) != 1 {
-		return "", nil
-	}
-	return candidates[0].Raw, nil
-}
-
-// IsUpgrade returns true if moving from currentRef to latestRef is a real
-// version upgrade. Returns false for noops where the current ref is already
-// at or more specific than the latest (e.g. v4.0.0 → v4, v3.1.1 → v3).
-func IsUpgrade(currentRef, latestRef string) bool {
-	if currentRef == latestRef {
-		return false
-	}
-	cur, curOK := parseSemver(currentRef)
-	lat, latOK := parseSemver(latestRef)
-	if !curOK || !latOK {
-		// If we can't parse the latest as semver, it's not a valid upgrade target.
-		if !latOK {
-			return false
-		}
-		// Can't parse current but latest is valid — assume it's an upgrade.
-		return true
-	}
-	// Skip pre-release in latest.
-	if lat.Rest != "" {
-		return false
-	}
-	// If latest major < current major, not an upgrade.
-	if lat.Major < cur.Major {
-		return false
-	}
-	// Same major — check if latest is just a less-specific version of current.
-	if lat.Major == cur.Major {
-		if lat.Minor == 0 && lat.Patch == 0 && cur.Minor >= 0 {
-			// latest=v4, current=v4.x.y → noop (major tag covers current)
-			// But also handle latest=v4 when the parsed minor/patch are 0
-			// because "v4" has no minor/patch components.
-			if latestRef == lat.MajorTag() {
-				return false
-			}
-		}
-		if lat.Minor == cur.Minor && lat.Patch <= cur.Patch {
-			return false
-		}
-		if lat.Minor < cur.Minor {
-			return false
-		}
-	}
-	return true
 }
 
 // RepoInfo holds repository metadata relevant for pinning decisions.
@@ -336,118 +134,6 @@ func (tl *TagLister) LookupTag(owner, repo, tagName string) *TagInfo {
 	return nil
 }
 
-// TagsForSHA returns all tags whose commit SHA matches the given SHA.
-func (tl *TagLister) TagsForSHA(owner, repo, sha string) ([]TagInfo, error) {
-	all, err := tl.ListTags(owner, repo)
-	if err != nil {
-		return nil, err
-	}
-	var matched []TagInfo
-	for _, t := range all {
-		if strings.EqualFold(t.SHA, sha) {
-			matched = append(matched, t)
-		}
-	}
-	return matched, nil
-}
-
-// SuggestTagsForSHA returns a curated set of tag suggestions for a pinned SHA.
-// It includes exact-match tags, plus major/minor family alternatives when the
-// match is version-like. Returns at most 5 suggestions.
-func (tl *TagLister) SuggestTagsForSHA(owner, repo, sha string) ([]TagSuggestion, error) {
-	matching, err := tl.TagsForSHA(owner, repo, sha)
-	if err != nil {
-		return nil, err
-	}
-
-	if len(matching) == 0 {
-		return nil, nil
-	}
-
-	var suggestions []TagSuggestion
-
-	// Find the best semver match to derive family tags.
-	var bestSV semver
-	bestFound := false
-	for _, t := range matching {
-		if sv, ok := parseSemver(t.Name); ok && sv.Rest == "" && !t.IsMajor {
-			if !bestFound || sv.Major > bestSV.Major ||
-				(sv.Major == bestSV.Major && sv.Minor > bestSV.Minor) ||
-				(sv.Major == bestSV.Major && sv.Minor == bestSV.Minor && sv.Patch > bestSV.Patch) {
-				bestSV = sv
-				bestFound = true
-			}
-		}
-	}
-
-	seen := make(map[string]bool)
-
-	// Add exact-match tags (non-major, non-minor-only) first.
-	for _, t := range matching {
-		if t.IsMajor {
-			continue
-		}
-		if len(suggestions) >= 3 {
-			break
-		}
-		suggestions = append(suggestions, TagSuggestion{
-			Tag:       t,
-			Reason:    "exact match for pinned SHA",
-			Preferred: true,
-		})
-		seen[t.Name] = true
-	}
-
-	// Suggest major/minor family tags if the best match is semver-ish.
-	if bestFound {
-		allTags, _ := tl.ListTags(owner, repo)
-
-		// Look for the major tag (e.g. v4).
-		majorName := bestSV.MajorTag()
-		if !seen[majorName] {
-			for _, t := range allTags {
-				if t.Name == majorName {
-					suggestions = append(suggestions, TagSuggestion{
-						Tag:    t,
-						Reason: fmt.Sprintf("major tag (tracks latest %s.x.x)", majorName),
-					})
-					seen[majorName] = true
-					break
-				}
-			}
-		}
-
-		// Look for a minor tag (e.g. v4.2) — not all repos have these.
-		minorName := bestSV.MinorTag()
-		if !seen[minorName] {
-			for _, t := range allTags {
-				if t.Name == minorName {
-					suggestions = append(suggestions, TagSuggestion{
-						Tag:    t,
-						Reason: fmt.Sprintf("minor tag (tracks latest %s.x)", minorName),
-					})
-					seen[minorName] = true
-					break
-				}
-			}
-		}
-	}
-
-	// Cap at 5 total.
-	if len(suggestions) > 5 {
-		suggestions = suggestions[:5]
-	}
-
-	return suggestions, nil
-}
-
-// PickerTag is a tag formatted for the interactive picker.
-type PickerTag struct {
-	Tag       TagInfo
-	Label     string // formatted display label
-	Installed bool   // true if this tag matches the currently pinned SHA
-}
-
 // LatestStableTag returns the latest non-major stable tag that passes cooldown.
 // It skips major-only tags (e.g. "v4"), pre-release tags, and tags younger
 // than the cooldown period. Returns ("", nil) if no suitable tag is found.
@@ -470,90 +156,6 @@ func (tl *TagLister) LatestStableTag(owner, repo string) (string, error) {
 		return t.Name, nil
 	}
 	return "", nil
-}
-
-// CuratePickerTags returns a short list of the most useful tags for a picker.
-// Shows the latest patch per major version (up to 3 majors), marks the one
-// matching pinnedSHA as "installed", and only puts 📦 on that one.
-func (tl *TagLister) CuratePickerTags(owner, repo, pinnedSHA string) ([]PickerTag, error) {
-	all, err := tl.ListTags(owner, repo)
-	if err != nil {
-		return nil, err
-	}
-	if len(all) == 0 {
-		return nil, nil
-	}
-
-	// Pick the latest full-version tag per major (skip major-only tags, pre-releases).
-	type majorBucket struct {
-		major int
-		tag   TagInfo
-	}
-	seen := make(map[int]bool)
-	var buckets []majorBucket
-
-	for _, t := range all {
-		if t.IsMajor {
-			continue
-		}
-		sv, ok := parseSemver(t.Name)
-		if !ok || sv.Rest != "" {
-			continue
-		}
-		// Skip tags younger than the cooldown period.
-		if tl.isTagTooNew(owner, repo, t.Name) && !strings.EqualFold(t.SHA, pinnedSHA) {
-			continue
-		}
-		if !seen[sv.Major] {
-			seen[sv.Major] = true
-			buckets = append(buckets, majorBucket{major: sv.Major, tag: t})
-		}
-		if len(buckets) >= 3 {
-			break
-		}
-	}
-
-	// Build picker entries.
-	var result []PickerTag
-	for _, b := range buckets {
-		installed := strings.EqualFold(b.tag.SHA, pinnedSHA)
-		result = append(result, PickerTag{
-			Tag:       b.tag,
-			Label:     b.tag.Name,
-			Installed: installed,
-		})
-	}
-
-	// If pinnedSHA matches a tag that isn't in our buckets, prepend it.
-	pinnedFound := false
-	for _, pt := range result {
-		if pt.Installed {
-			pinnedFound = true
-			break
-		}
-	}
-	if !pinnedFound {
-		for _, t := range all {
-			if strings.EqualFold(t.SHA, pinnedSHA) && !t.IsMajor {
-				label := t.Name + "  📦 installed"
-				result = append([]PickerTag{{
-					Tag:       t,
-					Label:     label,
-					Installed: true,
-				}}, result...)
-				break
-			}
-		}
-	}
-
-	return result, nil
-}
-
-// TagSuggestion is a tag paired with why it's being suggested.
-type TagSuggestion struct {
-	Tag       TagInfo
-	Reason    string
-	Preferred bool // true if this tag points directly at the pinned SHA
 }
 
 func (tl *TagLister) fetchTags(owner, repo string) ([]TagInfo, error) {
@@ -633,46 +235,6 @@ func (tl *TagLister) ReleaseDate(owner, repo, tag string) string {
 		return dates[tag]
 	}
 	return ""
-}
-
-// isMajorTag returns true if the tag looks like a major-only version (e.g. "v4", "v12").
-func isMajorTag(tag string) bool {
-	tag = strings.TrimPrefix(tag, "v")
-	for _, c := range tag {
-		if c < '0' || c > '9' {
-			return false
-		}
-	}
-	return len(tag) > 0
-}
-
-// FormatTagAge returns a relative age string like "3d ago" from an ISO 8601 timestamp.
-func FormatTagAge(isoDate string) string {
-	if isoDate == "" {
-		return ""
-	}
-	t, err := time.Parse(time.RFC3339, isoDate)
-	if err != nil {
-		return ""
-	}
-	d := time.Since(t)
-	switch {
-	case d < time.Hour:
-		return fmt.Sprintf("%dm ago", int(d.Minutes()))
-	case d < 24*time.Hour:
-		return fmt.Sprintf("%dh ago", int(d.Hours()))
-	case d < 30*24*time.Hour:
-		return fmt.Sprintf("%dd ago", int(d.Hours()/24))
-	case d < 365*24*time.Hour:
-		return fmt.Sprintf("%dmo ago", int(d.Hours()/(24*30)))
-	default:
-		return fmt.Sprintf("%dy ago", int(d.Hours()/(24*365)))
-	}
-}
-
-// TagURL returns a clickable GitHub releases tag URL.
-func TagURL(owner, repo, tag string) string {
-	return fmt.Sprintf("https://github.com/%s/%s/releases/tag/%s", owner, repo, tag)
 }
 
 // LoadCooldownConfig reads cooldown settings from ~/.config/gh-actions-pin/config.yml.
